Fixed musicmain.js not running
Accidentally deleted the p element with the 'musicmain' id that navigation.js looks for to detect whether the page is the music page or not, so musicmain.js didn't run whenever the page was accessed via local anchor. I'll probably change this so the page is constructed on the server instead of on the client in the future.